pyproject.toml
[tool.poetry]
name = "streamlit-app"
version = "0.1.0"
description = ""
authors = ["ManuelRucci <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
streamlit = "^1.7.0"
pandas = "^1.4.1"
plotly-express = "^0.4.1"
streamlit-webrtc = "^0.35.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
